requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
gunicorn==23.0.0
openai==1.97.0
PyPDF2==3.0.1
python-docx==1.2.0
psycopg2-binary==2.9.10
email-validator==2.2.0
SQLAlchemy==2.0.41
docx==0.2.4
python-dotenv
